New documentation site and examples app
Replaces the old storybook with static HTML documentation and an interactive
examples app. The documentation is now easier to customize, easier to maintain,
and faster to load. The interactive examples are a dedicated Next.js app that
can be loaded and edited in codesandbox, as well as embedded in the
documentation where needed.